backend/requirements.txt
bidict==0.23.1
blinker==1.7.0
bson==0.5.10
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
h11==0.14.0
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
six==1.16.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
cffi==1.16.0




